+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [.0.22.0] - 2020-05-24
+
 ### Added
 - New Debugger class and ability to trigger Debugger from module config screen (superusers only).